Reduce calls to malloc in unordered_multimap

+++ src/include/unordered_multimap.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2017-2019 Bailey Thompson
+ * Copyright (c) 2017-2020 Bailey Thompson
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -43,7 +43,7 @@ unordered_multimap_init(size_t key_size,
 
 /* Utility */
 int unordered_multimap_rehash(unordered_multimap me);
-int unordered_multimap_size(unordered_multimap me);
+size_t unordered_multimap_size(unordered_multimap me);
 int unordered_multimap_is_empty(unordered_multimap me);
 
 /* Accessing */
